refactor(SPV-1184): extract xpub, access key api test fixtures to separate files.
